file open macro
last bit of refactoring I am allowing myself to do for today.

it saves like 8 lines. In a 1000 line file. Sigh.

also, it addesses a symptom of a problem (distracting io code
in methods that do other things) without fixing the problem itself
(that said methods need to open files in the first place).

There is simply no low-hanging fruit to refactor here. It's all bad.